Create new function playRound
Function playRound which takes 2 parameters is created, it takes 2 inputs
and outputs who wins, the computer or the player. Paper beats rock, rock
beats scissors, scissors beat paper and if they choose the same, it's a
tie. A string is returned depending on the results of the game.

Earlier function computerChoice() was also changed to return the values
instead of console.log() them.