Create playGame() and update other functions
playGame() was created to play 5 games and a score will be calculated at
the end of the 5 games. Each win is 1 point, loss is 0 points and tie is
0.5 points. It will only be a tie if the score is 2.5 so playRound() was
updated to return points instead of the statements previously.

playRound() output statements are now using console.log instead of
return and the tie statement has been updated to return the capitalised
choice instead of the lower case version.